[CHANGE] Moved enums classes from `constants` to `app_settings`
To prevent potential circular imports.
