+++ aa_intel_tool/constants.py
@@ -4,7 +4,6 @@ App constants
 
 # Standard Library
 import re
-from enum import IntEnum
 
 # Django
 from django.utils.text import slugify
@@ -33,17 +32,6 @@ distance_units_off_grid: str = """
 """
 
 distance_units = f"{distance_units_on_grid}|{distance_units_off_grid}"
-
-
-class AdditionalEveCategoryId(IntEnum):
-    """
-    Eve category IDs which are not covered by Eve Universe
-    Unfortunately Python doesn't allow to extend eveuniverse.constants.EveCategoryId
-    """
-
-    DEPLOYABLE = 22
-    STARBASE = 23
-    SCANNER_PROBE = 479
 
 
 # Pre-compiled regex patterns used throughout the app
